+++ app.py
@@ -14,71 +14,56 @@
     the XMLRPC interface
 """
 
-import json
+import os
+import csv
+import time
+import logging
 import itertools
-from pprint import pprint
 
 from openpyxl import load_workbook
+from record import Record
 from api import API
 
-# Spreadsheet Stuff
-SPREADSHEET = '<your spreadsheet>.xlsx'
-SHEET = '<The Sheet name with the data>'
-FIRST_ROW = 2 # Assumes that the actual first row is a header
-LAST_ROW = 2064 # The last row that there is any data
-LAST_COL = 6 # The last column to read from in each row
-
-# Items in this list will not be checked for pre-existing records
-# and new records will always be created (at the risk of duplicate data)
-SERIALS_TO_IGNORE = []
-
 # Odoo Stuff
-ASSET_CATALOG_ID = 4525 # The database ID of the asset catalog we are importing into
-DATA_DESTRUCTION_ID = None # The database ID of the data destruction we are importing into
+# The database ID of the asset catalog we are importing into
+ASSET_CATALOG_ID = int(os.environ.get('odoo_asset_catalog_id', 0))
+# The database ID of the data destruction we are importing into
+DATA_DESTRUCTION_ID = int(os.environ.get('odoo_data_destruction_id', 0))
 
-class Record:
-    """
-        Stores information about the line items to import.
+# Spreadsheet Stuff
+SPREADSHEET = os.environ.get('spreadsheet', '')
+SHEET = os.environ.get('sheet', '')
+FIRST_ROW = int(os.environ.get('first_row', 1))
+LAST_ROW = int(os.environ.get('last_row', 2000))
+LAST_COL = int(os.environ.get('last_col', 6))
 
-        All fields but one are expected to be of type `str`,
-        and this class will enforce that when an instance is created.
-        The only field that is not expected to be a string is
-        the `children` field, which should be a list containing
-        one or more Record objects, or None
-    """
+# Items in this list will always create a Record (though that record
+# won't get uploaded to the ERP), and will additionally be added to
+# a spreadsheet as they are come across
+SERIALS_TO_IGNORE = os.environ.get('serials_to_ignore').strip().split('\n')
 
-    def __init__(self, **kwargs):
-        self.serial = str(kwargs.get('serial'))
-        self.asset_tag = str(kwargs.get('asset_tag'))
-        self.make = str(kwargs.get('make'))
-        self.model = str(kwargs.get('model'))
-        self.device_type = str(kwargs.get('device_type'))
+FILENAME_TIME = '%s' % (time.time())
+IGNORE_CSV = '%s.csv' % (FILENAME_TIME)
 
-        self.children = kwargs.get('children', list())
+# Logging
+logging.basicConfig(
+    filename='logs/%s.log' % (FILENAME_TIME),
+    level=logging.DEBUG,
+    format='%(asctime)s - %(levelname)s - %(message)s',
+    datefmt='%m/%d/%Y %I:%M:%S %p'
+)
+# Log to console and file
+CONSOLE = logging.StreamHandler()
+CONSOLE.setLevel(logging.INFO)
+FORMATTER = logging.Formatter('%(asctime)s %(levelname)s %(message)s')
+CONSOLE.setFormatter(FORMATTER)
+logging.getLogger('').addHandler(CONSOLE)
 
-    def __str__(self):
-        """
-            Returns the serial number of this Record instance
-        """
-        return self.serial
-
-    def __repr__(self):
-        """
-            Used when iterating over a list of Record objects
-
-            Returns a JSON serialized string that
-            represents this Record. Children will be
-            represented as a list of dictionaries or
-            `null`
-        """
-        return json.dumps({
-            'serial': self.serial,
-            'asset_tag': self.asset_tag,
-            'make': self.make,
-            'model': self.model,
-            'device_type': self.device_type,
-            'children': [child.__dict__ for child in self.children if child is not None],
-        })
+# Sanity Checks
+if ASSET_CATALOG_ID <= 0:
+    logging.warning('Zero or negative asset catalog. Records will not get uploaded to it')
+if DATA_DESTRUCTION_ID <= 0:
+    logging.warning('Zero or negative data destruction. Records will not get uploaded to it')
 
 class ProcessWorkbook:
     """
@@ -91,7 +76,22 @@ class ProcessWorkbook:
         self.api = API()
         self.workbook = load_workbook(filename=SPREADSHEET, data_only=True)[SHEET]
 
+        # Special Serials
+        self.ignore_csv_file = open(IGNORE_CSV, 'w')
+        self.ignore_csv = csv.DictWriter(
+            self.ignore_csv_file,
+            fieldnames=[
+                'serial', 'asset_tag', 'make',
+                'model', 'device_type', 'children'
+            ],
+            dialect=csv.excel
+        )
+        self.ignore_csv.writeheader()
+
+        # Failed records are rows that for one reason or another didn't generate a Record object
+        self.failed_records = list()
         self.records = list()
+        self.records_to_upload = list()
         self.last_parent = None
 
         # These lists are expected to contain one or more Tuples
@@ -106,7 +106,38 @@ class ProcessWorkbook:
         self.serials_to_ignore = [None, '', 'N/A']
         self.serials_to_ignore.extend(SERIALS_TO_IGNORE)
 
-        print('Initialized ProcessWorkbook')
+        self.rows_processed = 0
+        self.sorting_records_uploaded = 0
+        self.data_records_uploaded = 0
+        self.records_ignored = 0
+
+        logging.info('Initialized ProcessWorkbook')
+
+    def __del__(self):
+        """
+            Automatically closes file handlers when destructed normally
+        """
+        self.ignore_csv_file.close()
+
+        logging.info('Processed %d rows', (self.rows_processed))
+        logging.info('Created %d Records', (len(self.records)))
+        logging.info('Uploaded %d Sorting Assets', (self.sorting_records_uploaded))
+        logging.info('Uploaded %d Data Destruction Assets', (self.data_records_uploaded))
+        logging.info('Prevented %d Records from being uploaded', (self.records_ignored))
+
+        for row in self.failed_records:
+            # pylint: disable=logging-not-lazy
+            logging.info(
+                'Row that failed Record Creation: %s | %s | %s | %s | %s' % (
+                    row[0].value,
+                    row[1].value,
+                    row[3].value,
+                    row[4].value,
+                    row[5].value
+                )
+            )
+
+        logging.info('ProcessWorkbook Finished')
 
     def get_id_from_model(self, model):
         """
@@ -132,7 +163,11 @@ class ProcessWorkbook:
             `records` is an optional argument that will
             search that specific record list (such as a
             parent's children) instead of just the parent
-            records
+            records.
+
+            If the serial is to be ignored, this will
+            always return False, to allow the Record object
+            to be created (used to save it to a csv)
         """
         if records is None:
             records = self.records
@@ -146,7 +181,9 @@ class ProcessWorkbook:
         """
             With a provided `row`, this method will first search
             for a matching serial number and if it doesn't exist,
-            a new Record will be created with that row's data
+            a new Record will be created with that row's data.
+            If the row's serial number is special, the record will
+            be created regardless of one already existing.
 
             `parent` is an optional argument that sets the `last_parent`
             attribute. `last_parent` is used to append new Records to
@@ -159,10 +196,13 @@ class ProcessWorkbook:
             If the serial number isn't in the list, this returns the
             created Record object. Otherwise, it returns False
         """
-        if not self.serial_in_records(str(row[0].value)):
+
+        serial = str(row[0].value)
+        if not self.serial_in_records(serial):
             # pylint: disable=bad-whitespace
+            logging.debug('Creating Record for %s', (serial))
             record = Record(
-                serial = str(row[0].value),
+                serial = serial,
                 asset_tag = str(row[1].value),
                 make = str(row[3].value),
                 model = str(row[4].value),
@@ -201,7 +241,7 @@ class ProcessWorkbook:
 
             Returns `self` (this instance of ProcessWorkbook)
         """
-        print('Getting rows from the spreadsheet and sorting relationships')
+        logging.info('Getting rows from the spreadsheet and sorting relationships')
         # pylint: disable=bad-continuation
         for row in self.workbook.iter_rows(
             min_row=FIRST_ROW,
@@ -226,23 +266,27 @@ class ProcessWorkbook:
                 if record:
                     self.records.append(record)
 
+            if not record:
+                self.failed_records.append(row)
+
+            self.rows_processed += 1
+
         return self
 
     def get_records(self):
         """
             Deprecated, use `build_record_list` instead.
         """
-        print('get_records is deprecated and will be removed in the future.')
-        print('Please use `build_record_list` instead.')
+        logging.warning('get_records is deprecated and will be removed in the future.')
+        logging.warning('Please use `build_record_list` instead.')
         return self.build_record_list()
 
     def show_records(self):
         """
-            Pretty Prints a JSON string for all
-            of the records that are stored
+            Logs all of the records stored, in JSON format
         """
         # pylint: disable=unnecessary-comprehension
-        pprint([record for record in self.records])
+        logging.debug([record for record in self.records])
 
     def get_odoo_model_ids(self):
         """
@@ -252,17 +296,17 @@ class ProcessWorkbook:
             contains a mapping between each unique model
             name and the database id.
 
-            When a model can't be found, that model is printed
+            When a model can't be found, that model is logged
             so that a manual search can be done, or a new item
             can be created.
 
             When a model returns multiple ids, that model is
-            printed so that a manual search can be done to
+            logged so that a manual search can be done to
             select the "correct" database id.
 
             Returns `self` (this instance of ProcessWorkbook)
         """
-        print('Searching Odoo for sellable items with matching models')
+        logging.info('Searching Odoo for sellable items with matching models')
         for model in self.models_to_search:
             odoo_records = self.api.do_search_and_read(
                 'erpwarehouse.sellable',
@@ -270,7 +314,7 @@ class ProcessWorkbook:
             )
 
             if not odoo_records:
-                print('Unable to find model: %s' % model[1])
+                logging.warning('Unable to find model: %s', (model[1]))
                 self.models_to_create.append(model)
             else:
                 # If the search returns more than one, we'll
@@ -286,9 +330,9 @@ class ProcessWorkbook:
 
             Returns `self` (this instance of ProcessWorkbook)
         """
-        print('Creating sellable items for missing models')
+        logging.info('Creating sellable items for missing models')
         for model in self.models_to_create:
-            print('Creating model: %s' % model[1])
+            logging.info('Creating model: %s', (model[1]))
             result = self.api.do_create(
                 'erpwarehouse.sellable',
                 {
@@ -300,24 +344,54 @@ class ProcessWorkbook:
 
         return self
 
+    def asset_line_exists(self, record):
+        """
+            Searches the asset catalog for
+            a line item matching the provided `record`.
+
+            Determines if a line is the same if either the
+            serial number was previously recorded.
+
+            Returns True if there is an existing record in Odoo,
+            False otherwise.
+        """
+        logging.debug('Checking if "%s" already exists before creation in Odoo', (record.serial))
+        result = self.api.do_search(
+            'erpwarehouse.asset',
+            [
+                ('catalog', '=', ASSET_CATALOG_ID),
+                ('make', '=', self.get_id_from_model(record.model)),
+                ('serial', '=ilike', record.serial),
+            ]
+        )
+        if len(result) > 0:
+            return True
+        return False
+
     def _create_asset_catalog_line(self, record):
         """
             With the provided `record` (Record) instance,
             this method will issue an API request to Odoo
-            to create the line item
+            to create the line item after searching Odoo
+            for that record.
         """
         if self.get_id_from_model(record.model):
-            result = self.api.do_create(
-                'erpwarehouse.asset',
-                {
-                    'catalog': ASSET_CATALOG_ID,
-                    'make': self.get_id_from_model(record.model),
-                    'serial': record.serial,
-                    'tag': record.asset_tag,
-                })
-            print('Added id: %s' % (result))
+            if not self.asset_line_exists(record):
+                result = self.api.do_create(
+                    'erpwarehouse.asset',
+                    {
+                        'catalog': ASSET_CATALOG_ID,
+                        'make': self.get_id_from_model(record.model),
+                        'serial': record.serial,
+                        'tag': record.asset_tag,
+                    }
+                )
+                self.sorting_records_uploaded += 1
+                logging.debug('Added id: %s', (result))
+            else:
+                logging.warning('"%s" already existed, so it was skipped', (record.serial))
         else:
-            print('Unable to add %s as there is no sellable id' % (record.serial))
+            logging.error('Unable to add "%s" as there is no sellable id', (record.serial))
 
         return self
 
@@ -354,9 +428,10 @@ class ProcessWorkbook:
                     'storser': child.serial if child else 'N/A',
                     'type': device_type,
                 })
-            print('Added id: %s' % (result))
+            self.data_records_uploaded += 1
+            logging.debug('Added id: %s', (result))
         else:
-            print('Unable to add %s as there is no sellable id' % (record.serial))
+            logging.error('Unable to add "%s" as there is no sellable id', (record.serial))
 
         return self
 
@@ -369,8 +444,8 @@ class ProcessWorkbook:
 
             Returns `self` (this instance of ProcessWorkbook)
         """
-        print('Creating Line items for accepted records in Odoo')
-        for record in self.records:
+        logging.info('Creating Line items for accepted records in Odoo')
+        for record in self.records_to_upload:
 
             if ASSET_CATALOG_ID:
                 self._create_asset_catalog_line(record)
@@ -384,13 +459,45 @@ class ProcessWorkbook:
 
         return self
 
+    def remove_ignored_records(self):
+        """
+            Populates `self.records_to_upload` with
+            any record that is not also a part of the
+            `self.serials_to_ignore` list, counts the
+            serials that were ignored, and writes those
+            rows to an ignore csv
+        """
+        logging.info('Removing Ignored Serials from Records')
+
+        self.records_to_upload = [
+            x for x in self.records
+            if x.serial not in self.serials_to_ignore
+        ]
+
+        for record in self.records:
+            if record.serial in self.serials_to_ignore:
+                self.records_ignored += 1
+                logging.warning(
+                    '"%s" is special, skipping import and saving to special list', (record.serial)
+                )
+                self.ignore_csv.writerow({
+                    'serial': record.serial,
+                    'asset_tag': record.asset_tag,
+                    'make': record.make,
+                    'model': record.model,
+                    'device_type': record.device_type,
+                    'children': record.children,
+                })
+
     def run(self):
         """
             Runs everything in the order that is required
         """
         self.build_record_list()
+        self.show_records()
         self.get_odoo_model_ids()
         self.create_missing_model_ids()
+        self.remove_ignored_records()
         self.create_line_items()
 
 if __name__ == '__main__':