+++ changelog.md
@@ -7,7 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.1.1] - 2020-04-06
+## [1.2.0] - 2020-05-28
+
+### Added
+
+- Saving of "Serials to Ignore" to a CSV file. This is used to select lines that shouldn't be imported, but do need to be looked at carefully
+- The number of rows being processed is kept track of and reported at the end of the program
+- CSV files generated as a result of this program are now ignored in git
+- We now keep track of the number of each type of record uploaded
+- We now track instances of failed records and see the data. Normally when a record fails to be created, it is due to being a duplicate.
+
+### Changed
+
+- Asset lines are now checked to see if they exist in Odoo before blindly creating new records (Asset Catalog only. This is not currently on the data destruction)
+- Record objects are now able to be duplicated for cases of "Serials to Ignore" (Multiple Records with the same "Serial Number")
+- The Print statements have been replaced with native logging, which output to STDOUT and a file
+- Configuration of this program has been moved to `run.sh`. You should not need to alter `app.py` going forward.
+- The Record class has been broken out into its own file, to help with readability
+
+## [1.1.1] - 2020-05-27
 
 ### Added